La valoración entera intervalar se ha cambiado. El max y el min ahora son double porque al normalizar la división salía entera. Estamos cambiando la forma de calcular los pesos en el TODIM. Ahora se cogen del operador de agregación. El cálculo del peso de los criterios en el Fuzzy TODIM me genera dudas

+++ sinbad2.aggregationoperator.median/src/sinbad2/aggregationoperator/median/valuation/IntervalIntegerOperator.java
@@ -39,8 +39,8 @@ public class IntervalIntegerOperator {
 			
 			if((values.size() % 2) == 0) {
 				Valuation auxValuation = values.get(median - 1);
-				long minValue = (((IntegerIntervalValuation) result).getMin()) + (((IntegerIntervalValuation) auxValuation).getMin());
-				long maxValue = (((IntegerIntervalValuation) result).getMax()) + (((IntegerIntervalValuation) auxValuation).getMax());
+				double minValue = (((IntegerIntervalValuation) result).getMin()) + (((IntegerIntervalValuation) auxValuation).getMin());
+				double maxValue = (((IntegerIntervalValuation) result).getMax()) + (((IntegerIntervalValuation) auxValuation).getMax());
 				minValue /= 2;
 				maxValue /= 2;
 				((IntegerIntervalValuation) result).setMinMax(minValue, maxValue);
